+++ apps/linksense-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # linksense-react
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @linksense/react@1.0.0
+  - @linksense/core@1.0.0
+
 ## 0.0.1
 
 ### Patch Changes