+++ packages/react/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @linksense/react
 
+## 1.0.0
+
+### Major Changes
+
+- all adapter with ai support added
+
+### Patch Changes
+
+- Updated dependencies
+  - @linksense/core@1.0.0
+
 ## 0.1.1
 
 ### Patch Changes